requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
automium==0.2.6
automium-web==0.1.1
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.6
click==7.1.2
cryptography==3.4.8
distlib==0.3.1
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.1
filelock==3.0.12
flake8==3.9.2
Flask==1.1.2
h11==0.12.0
idna==3.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
passlib==1.7.4
pbr==5.6.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyJWT==2.1.0
pymongo==3.12.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.5
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
testresources==2.0.1
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
virtualenv==20.0.27
Werkzeug==1.0.1
